Proof of concept, rough, single file processor
Takes the sample file and reworks the file layout into two different timeseries dataframes.  One dataframe consists of the wave and ocean data and the other is the wave spectra data.

Each data file can store multiple days worth of data, each field is output in a specific order on an individual line.  Records are separated by 2 blank lines, the wave spectra data is actually comprised of 4 "columns" separated by a semi-colon at the end of each record after the more standard timeseries data.

While logical, this format is not compatible with ERDDAP as is and needs to be reformatted - given that each record of wave data is accompanied with 64 records of wave spectra it may be best to split these files into 2 separate datasets.