Added configuration options for output to CSV
Added configuration options for dumping PVM data into CSV format.  2 Datasets are created in this way, timeseries wave data and wave spectra.

The timeseries data will be split into 1 file/day, which means existing files and previous data must be accounted for.  Differences in date/time indexes are searched for, if the new PVM file has date/time indexes that are not present in the existing file, the 2 dataframes are concatenated with the existing data being preserved and the new records added.

In the case of the wave spectra data, the many records associated with a single timestamp makes it much simpler to dump each observation into its own 64 row file.  The volume of files created this way means a deeper directory structure is desirable since a single year could create 17.520 files if sampled every half hour per station.