Updated to process one or many files
The batch flag will accept a wildcard argument to select multiple files starting from path specified in the source_file argument.

The JSON configuration file is loaded once at the beginning of execution and refreshed on each iteration of the loop when batch processing - this keeps the internal pandas series and dataframes clean instead of constantly building up.